for compile

+++ src/service.cpp
@@ -19,6 +19,7 @@
 #include "log.h"
 #include "service.h"
 #include "configure.h"
+#include <stdio.h>
 
 int Service::init(const char *path, const char *file)
 {
@@ -148,7 +149,7 @@ int Service::get_connection(Connection &conn)
     for (int n = 0; n < 3 && avail_server_num > 0; ++n) /* retry */
     {
         int idx = conn._key % avail_server_num;
-        for (size_t i = 0; i < _server_num; ++i)
+        for (int i = 0; i < _server_num; ++i)
         {
             if (flags[i])
             {
